requirements.txt
aiogram~=3.21.0
python-dotenv~=1.1.1
httpx~=0.28.1